Introduce separate AMP & Jetpack integration classes
