+++ includes/Plugin.php
@@ -28,6 +28,8 @@
 
 namespace Google\Web_Stories;
 
+use Google\Web_Stories\Integrations\AMP;
+use Google\Web_Stories\Integrations\Jetpack;
 use Google\Web_Stories\Integrations\Site_Kit;
 use Google\Web_Stories\REST_API\Embed_Controller;
 use Google\Web_Stories\REST_API\Stories_Media_Controller;
@@ -141,6 +143,20 @@ class Plugin {
 	public $experiments;
 
 	/**
+	 * AMP plugin integration.
+	 *
+	 * @var AMP
+	 */
+	public $amp;
+
+	/**
+	 * Jetpack integration.
+	 *
+	 * @var Jetpack
+	 */
+	public $jetpack;
+
+	/**
 	 * Site Kit integration.
 	 *
 	 * @var Site_Kit
@@ -217,6 +233,12 @@ class Plugin {
 
 		$activation_notice = new Activation_Notice( $activation_flag );
 		$activation_notice->init();
+
+		$this->amp = new AMP();
+		add_action( 'init', [ $this->amp, 'init' ] );
+
+		$this->jetpack = new Jetpack();
+		add_action( 'init', [ $this->jetpack, 'init' ] );
 
 		$this->site_kit = new Site_Kit( $this->analytics );
 		add_action( 'init', [ $this->site_kit, 'init' ] );
